RF-DETR runs 2-4: final 700 px checkpoint, training fixes, YOLO setup
Checkpoints (Git LFS):
  runs/rfdetr_run2/best.pt      560 px, epoch 56 of 60 (converged)
                                test mAP@50 0.460 / mAP@50-95 0.258
  runs/rfdetr_run4_700px/best.pt 700 px, epoch 19 of 40 (final model)
                                test mAP@50 0.472 / mAP@50-95 0.264
Runs were warm-started in a chain: run1 ep31 -> run2 -> run3 (700 px,
20 ep) -> run4. 130 epochs total. Logs, configs, test/val per-class
evaluations and regenerated figures for all runs are included.

Training fixes:
  - detr_criterion: Hungarian matching batched across images and decoder
    layers (one GPU->CPU sync per step instead of 56). Losses and
    assignments verified bit-identical; criterion 3x faster.
  - evaluate: val_precision was a copy of val_mAP50 (averaged the whole
    PR curve). Now precision at each class's F1-optimal operating point.
  - autoconfig: 8 GB tier retuned from measurements (batch 8 x accum 2,
    4 workers); resolution snapped to the model's grid (14 / 32).
  - dataset: reduced-size JPEG decoding via PIL draft() for the 8 MP
    TACO photos; boxes identical, loader 57 -> 132 img/s.
  - train_loop: --resume from last.pt (raw + EMA weights, optimizer,
    schedule position, best-so-far); last.pt now carries full state.
  - train: --init-from, --accum, --workers; keep-awake against idle
    sleep; validation loader uses fewer non-persistent workers (RAM).
  - yolo_loss: loss/assignment in fp32 (crashed on first batch under
    bf16 autocast).

New: scripts/predict.py (inference on HEIC/JPEG with drawn boxes),
train_yolo.bat (YOLOv11-s baseline launcher, 576 px, auto-resume).

+++ civicscan_training_rfdetr/training/civicscan/models/yolo_loss.py
@@ -125,8 +125,13 @@ class YOLOLoss(nn.Module):
         return loss.view(-1, 4).mean(1)
 
     def forward(self, outputs, targets, valid_mask):
-        cls_logits = outputs["pred_cls"]                    # [B,A,C]
-        dist = outputs["pred_dist"]                         # [B,A,4*reg_max]
+        # Loss and assignment run in fp32 regardless of the autocast dtype of
+        # the forward pass (the model itself stays bf16/fp16). Under bf16 the
+        # assigner otherwise builds target_scores in bf16 and index-puts fp32
+        # alignment values into it, which raises on the first batch; and the
+        # TAL power/IoU arithmetic is not something to do in 8-bit mantissas.
+        cls_logits = outputs["pred_cls"].float()            # [B,A,C]
+        dist = outputs["pred_dist"].float()                 # [B,A,4*reg_max]
         anchors, strides = outputs["anchors"], outputs["strides"]
         b, a, c = cls_logits.shape
         dev = cls_logits.device
